Refactor docstring detection in separate files (code & tests)
