+++ test/test_checks.py
@@ -1,7 +1,3 @@
-import tokenize
-
-from flake8_quotes import Token, get_docstring_tokens
-
 from flake8_quotes import QuoteChecker
 import os
 import subprocess
@@ -231,31 +227,6 @@ class DocstringTestChecks(TestCase):
             {'col': 4, 'line': 17, 'message': 'Q001 Remove bad quotes from multiline string.'},
             {'col': 8, 'line': 29, 'message': 'Q001 Remove bad quotes from multiline string.'},
         ])
-
-
-class GetDocstringTokensTests(TestCase):
-    def _get_docstring_tokens(self, filename):
-        f = open(get_absolute_path(filename), 'r')
-        tokens = [Token(t) for t in tokenize.generate_tokens(f.readline)]
-        return get_docstring_tokens(tokens)
-
-    def test_get_docstring_tokens_absent(self):
-        self.assertEqual(self._get_docstring_tokens('data/doubles.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/doubles_multiline_string.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/doubles_noqa.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/doubles_wrapped.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/multiline_string.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/no_qa.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/singles.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/singles_multiline_string.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/singles_noqa.py'), set())
-        self.assertEqual(self._get_docstring_tokens('data/singles_wrapped.py'), set())
-
-    def test_get_docstring_tokens(self):
-        f = open(get_absolute_path('data/docstring_doubles.py'), 'r')
-        tokens = [Token(t) for t in tokenize.generate_tokens(f.readline)]
-        docstring_tokens = get_docstring_tokens(tokens)
-        self.assertEqual(docstring_tokens, {tokens[1], tokens[18], tokens[45]})
 
 
 def get_absolute_path(filepath):
